# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API for Etchebest video streaming platform",
        "title": "Etchebest API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/api/discover": {
            "get": {
                "description": "Returns trending country, top dishes and categories",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Discover"
                ],
                "summary": "Discover home",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/discover/categories": {
            "get": {
                "description": "Returns all active countries with live count",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Discover"
                ],
                "summary": "List discover categories",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/discover/categories/{id}/lives": {
            "get": {
                "description": "Returns paginated lives for a specific country with filters",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Discover"
                ],
                "summary": "Get lives by category",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Category ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Page size",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Sort by (views, viewers, recent)",
                        "name": "sort",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Live status",
                        "name": "filters[status]",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Dish ID",
                        "name": "filters[dish_id]",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Minimum viewers",
                        "name": "filters[min_viewers]",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Lives started in last N hours",
                        "name": "filters[hours_ago]",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}